apps/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
sqlmodel==0.0.14
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1
redis==5.0.1
websockets==12.0
twilio==8.10.0
python-dotenv==1.0.0
aiofiles==23.2.1
slowapi==0.1.9
razorpay==1.4.1
boto3==1.34.10
cryptography==41.0.7
livekit-api==0.6.4
livekit==0.11.1
httpx==0.27.0
